assignments/assignment_09/pyproject.toml
[project]
name = "assignment-09"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
